+++ src/fx/explode.rs
@@ -113,14 +113,16 @@ impl Shader for Explode {
         let cell_filter = self
             .cell_filter
             .as_ref()
-            .map(|f| f.predicate(safe_area))
-            .unwrap_or(CellFilter::All.predicate(safe_area));
+            .map(FilterProcessor::validator);
 
         for (pos, (dx, dy)) in cells.iter() {
             let pos = *pos;
             let (dx, dy) = (*dx, *dy);
 
-            if !cell_filter.is_valid(pos, &buf[pos]) {
+            if cell_filter
+                .as_ref()
+                .is_some_and(|c| !c.is_valid(pos, &buf[pos]))
+            {
                 continue;
             }